RGB LED Experiments
Created the circuit which runs an RGB LED, determined in the design document how the RGB LED should be connected and the strange requirements I discovered to make it work correctly. (1000ohm on green, 330ohm on red and blue).

The code here has been designed so to be modularly included in the final project. Following the state based design there in.